Extract WhatsNewHandler
